Revert "Merge branch 'Experiment' into MainClone"
This reverts commit 76199e1582c53eef600de84349b166441070ebf0, reversing
changes made to dfb85ac0e4b8cd1d7bfe0dcbe2aface0f78f8593.

+++ Assets/Scripts/Entity/StatusEffect/Buff/Heal.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class Heal : StatusEffect
 {
     public float CasterATK;
-    public float percent;
+    public int percent;
 
     public override void OnApply(CharacterInBattle target)
     {
