+++ Assets/Scripts/Entity/StatusEffect/Buff/SpeedUP.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewSpeedUP", menuName = "StatusEffect/SpeedUP")]
 public class SpeedUP : StatusEffect
 {
-    public float PercentAmount;
+    public int PercentAmount;
 
     public override void OnApply(CharacterInBattle target)
     {
